backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
psycopg2-binary==2.9.9
asyncpg==0.29.0
sqlalchemy==2.0.23
alembic==1.13.0

# Supabase
supabase==2.10.0
email-validator==2.1.0

# Redis & Caching
redis==5.0.1
celery==5.3.4

# API & Validation
requests==2.31.0
httpx==0.25.2

# Data Processing
pandas==2.1.3
numpy==1.26.3

# ML & AI
tensorflow==2.15.0
scikit-learn==1.3.2
transformers==4.35.2

# Technical Indicators
ta==0.10.2
pandas-ta==0.3.14b

# Async & Performance
aiohttp==3.9.1
websockets==12.0

# Monitoring & Logging
prometheus-client==0.19.0
python-json-logger==2.0.7
structlog==23.3.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Utils
python-dateutil==2.8.2
pytz==2023.3
croniter==2.0.1

# Development
black==23.12.0
flake8==6.1.0
isort==5.13.2

# Exchange APIs
ccxt==4.0.94
